Created no internet popup
